+++ test/e2e/config_test.go
@@ -118,6 +118,7 @@ func TestClusterMonitorPrometheusK8Config(t *testing.T) {
 	data := fmt.Sprintf(`prometheusK8s:
   logLevel: debug
   retention: 10h
+  retentionSize: 15GB
   queryLogFile: /tmp/test.log
   tolerations:
     - operator: "Exists"
@@ -156,6 +157,7 @@ func TestClusterMonitorPrometheusK8Config(t *testing.T) {
 					expectMatchingRequests(podName, containerName, mem, cpu),
 					expectContainerArg("--log.level=debug", containerName),
 					expectContainerArg("--storage.tsdb.retention.time=10h", containerName),
+					expectContainerArg("--storage.tsdb.retention.size=15GB", containerName),
 				},
 			),
 		},